+++ test/controls/Phone.d.ts
@@ -1,0 +1,6 @@
+export declare let data: {
+    list: never[];
+};
+export declare let methods: {
+    action: (this: any, name: string, item: string[], index: number) => void;
+};